Only render bytevectors as text if all characters are printable. (#4561)
Previously bytevectors were logged as text as long as they didn't contain null bytes. This could cause crashes if verbose logging in the enclave was enabled, the host used the JSON logger and the application wrote binary data to the KV. The log messages would contain invalid Unicode sequences, which made the JSON logger crash when trying to embed the message as a string field.